feat: new settings
